+++ src/mruby_integration/textures.cpp
@@ -34,30 +34,6 @@ mrb_load_texture(mrb_state* mrb, mrb_value) -> mrb_value
 }
 
 auto
-mrb_load_texture_from_image(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  Image* image;
-  mrb_get_args(mrb, "d", &image, &Image_type);
-
-  auto* texture = static_cast<Texture2D*>(malloc(sizeof(Texture2D)));
-  *texture = LoadTextureFromImage(*image);
-
-  mrb_value obj = mrb_obj_value(
-    Data_Wrap_Struct(mrb, Texture2D_class, &Texture2D_type, texture));
-
-  setup_Texture2D(mrb,
-                  obj,
-                  texture,
-                  texture->id,
-                  texture->width,
-                  texture->height,
-                  texture->mipmaps,
-                  texture->format);
-
-  return obj;
-}
-
-auto
 mrb_unload_texture(mrb_state* mrb, mrb_value) -> mrb_value
 {
   Texture2D* texture;
@@ -165,11 +141,6 @@ append_textures(mrb_state* mrb)
 {
   mrb_define_method(
     mrb, mrb->kernel_module, "load_texture", mrb_load_texture, MRB_ARGS_REQ(1));
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "load_texture_from_image",
-                    mrb_load_texture_from_image,
-                    MRB_ARGS_REQ(1));
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "unload_texture",